Fix top navigation
