+++ client/components/about.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import TopNav from './topNav';
+import AuthModal from './authModal';
 
 export default class About extends React.Component {
   constructor(props) {
@@ -21,6 +22,22 @@ export default class About extends React.Component {
   }
 
   componentWillMount() {
+    // $.ajax({
+    //   type: 'GET',
+    //   url: '/verifyAuth',
+    //   async: false,
+    //   success: function(data) {
+    //     if (!data.success) {
+    //         console.log('auth fail')
+    //         $('#nav-profile').css('display', 'none');
+    //         $('#nav-auth').css('display', 'block')
+    //     } else {
+    //         console.log('auth success')
+    //         $('#nav-profile').css('display', 'block');
+    //         $('#nav-auth').css('display', 'none')
+    //     }
+    //   }
+    // })
     this.randomizeCreatorsOrder();
     this.props.router.setRouteLeaveHook(
         this.props.route,
@@ -65,9 +82,10 @@ export default class About extends React.Component {
     return(
     <div id="aboutPage">
       <TopNav router={this.props.router}/>
+      <AuthModal />
       <div className="aboutPageContent">
         <h1 className="headers">The Game</h1>
-        <p>VR Pacman is a spin on the classic 1980s Pac-man game by Namco. In VR Pacman, you become the Pac-man in this first-person maze traversing game. Users can play in different game modes, as well as compete online or play private games with their friends. They can also design or purchase their own virtual reality mazes and play in them. This game is meant for both virtual reality and PC.</p>
+        <p>VR Pacman is a spin on the classic 1980s Pac-man game by Namco. In VR Pacman, you become the Pac-man in this first-person maze traversing game. Users can play in different game modes, competing in ranked games or playing custom private games with their friends. They can also design or purchase their own virtual reality mazes and share them to the world. This game is meant for both virtual reality and PC.</p>
         <p>The game was built using the BabylonJS library on top of a React, Node, Express, MySQL/Sequelize stack.</p>
         <h1 className="headers">The Creators</h1>
         <p>We are a team of 3 software engineers excited about VR and web development.</p>
