+++ client/components/profilePage.jsx
@@ -20,32 +20,7 @@ export default class ProfilePage extends React.Component {
     this.convertData = this.convertData.bind(this);
   }
 
-  logout() {
-    $.ajax({
-      type: 'GET',
-      url: '/logout',
-      success: function() {
-        localStorage.clear();
-        this.props.router.push({pathname: '/'});
-      }.bind(this)
-    })
-  }
-
   componentWillMount() {
-    $.ajax({
-      type: 'GET',
-      url: 'profileInfo',
-      async: false,
-      success: function(data) {
-        console.log('profile info data', data);
-        this.setState({
-          username: data.username,
-          spHighScore: data.spHighScores_PC,
-          mpHighScore: data.mpHighScores_PC
-        });
-        localStorage.setItem('username', data.username);
-      }.bind(this)
-    });
     this.getMyMazes();
     window.selectedMaze = null;
   }
@@ -239,12 +214,6 @@ export default class ProfilePage extends React.Component {
     return (
       <div>
         <TopNav router={this.props.router}/>
-        <div id="profileStats">
-          <p id="welcomeMessage">Welcome {this.state.username}</p>
-          <p>SP High Score: {this.state.spHighScore}</p>
-          <p>MP High Score: {this.state.mpHighScore}</p>
-          <button id="logout" onClick={this.logout.bind(this)}>Log Out</button>
-        </div>
         <div className="profilePageContent">
           <div className="playScreen">
             <h1 className="headers">Play</h1>
